Fix in case libtool doesn't have a timestamp
git-svn-id: svn://localhost/CW-AutoMacros.svn,repos,svnroot,svn2.berlios.de/trunk@60 6a1ecc81-c4c8-904c-b5a8-411a816c4db3

+++ scripts/autogen.sh
@@ -290,8 +290,8 @@ if test "$using_libtool" = "yes"; then
   # Check if bootstrap was run before and if the installed files are the same version.
   if test -f ltmain.sh; then
     installed_libtool=`grep '^VERSION=' ltmain.sh | sed -r -e 's/([^12]|[12][^.])*([12]\.[0-9]+(\.[0-9]+)*).*/\2/'`
-    installed_timestamp=`grep '^TIMESTAMP=' ltmain.sh | sed -e 's/.*(\([0-9]*\.[^ ]*\).*/\1/'`
-    if test "$installed_libtool" != "$libtool_version" -o X"$installed_timestamp" != X"$libtool_develversion"; then
+    installed_timestamp=`grep '^TIMESTAMP=' ltmain.sh | sed -e 's/.*(\([0-9]*\.[^ ]*\).*/\1/;s/TIMESTAMP=""/no timestamp/'`
+    if test "$installed_libtool" != "$libtool_version" -o \( X"$installed_timestamp" != X"$libtool_develversion" -a X"$installed_timestamp" != X"no timestamp" \); then
       echo "Re-installing new libtool files ($installed_libtool ($installed_timestamp) -> $libtool_version ($libtool_develversion))"
       rm -f config.guess config.sub ltmain.sh ltconfig
     fi
